requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.3
psycopg2-binary
requests>=2.31.0
Pillow>=10.0.0
google-auth>=2.29.0

